Extracted popup menu function to Generic
