+++ app/src/main/res/values/strings.xml
@@ -66,5 +66,6 @@
     <string name="watched_tab">Watched</string>
     <string name="movie_tag">Movie</string>
     <string name="show_tag">Show</string>
-    <string name="remove_options_popup_menu">Remove from Watchlist</string>
+    <string name="remove_option_popup_menu">Remove from %s</string>
+    <string name="add_option_popup_menu">Add to %s</string>
 </resources>
